use of authorization by API key added

+++ src/TaskGateway.php
@@ -11,33 +11,38 @@ class TaskGateway
         $this->connection = $database->getConnection();
     }
 
-    public function all(): array
+    public function all(int $userId): array
     {
-        $sql = sprintf('SELECT * FROM %s', self::TASKS_TABLE);
+        $sql = sprintf('SELECT * FROM %s WHERE user_id = :user_id ORDER BY name', self::TASKS_TABLE);
 
-        $stmt = $this->connection->query($sql);
+        $stmt = $this->connection->prepare($sql);
+
+        $stmt->bindValue(':user_id', $userId, PDO::PARAM_INT);
+
+        $stmt->execute();
 
         return $stmt->fetchAll(PDO::FETCH_ASSOC);
     }
 
-    public function get(string $id): array|false
+    public function get(int $userId, string $id): array|false
     {
-        $sql = sprintf('SELECT * FROM %s WHERE id = :id', self::TASKS_TABLE);
+        $sql = sprintf('SELECT * FROM %s WHERE id = :id AND user_id = :user_id', self::TASKS_TABLE);
 
         $stmt = $this->connection->prepare($sql);
 
         $stmt->bindValue(':id', $id, PDO::PARAM_INT);
+        $stmt->bindValue(':user_id', $userId, PDO::PARAM_INT);
 
         $stmt->execute();
 
         return $stmt->fetch(PDO::FETCH_ASSOC);
     }
 
-    public function create(array $data): string
+    public function create(int $userId, array $data): string
     {
         $sql = sprintf(
-            'INSERT INTO %s (name, priority, is_completed) 
-            VaLUES (:name, :priority, :is_completed)', 
+            'INSERT INTO %s (name, priority, is_completed, user_id) 
+            VaLUES (:name, :priority, :is_completed, :user_id)', 
             self::TASKS_TABLE
         );
 
@@ -50,13 +55,14 @@ class TaskGateway
             empty($data['priority']) ? PDO::PARAM_NULL : PDO::PARAM_INT
         );
         $stmt->bindValue(':is_completed', $data['is_completed'] ?? false, PDO::PARAM_BOOL);
+        $stmt->bindValue(':user_id', $userId, PDO::PARAM_INT);
 
         $stmt->execute();
 
         return $this->connection->lastInsertId();
     }
 
-    public function update(string $id, array $data): int
+    public function update(int $userId, string $id, array $data): int
     {
         $fields = [];
 
@@ -83,15 +89,18 @@ class TaskGateway
                 array_keys($fields)
             );
 
-            $sql = sprintf('UPDATE %s SET ' . implode (', ', $sets) . ' WHERE id = :id', self::TASKS_TABLE);
+            $sql = sprintf('UPDATE %s SET ' . implode (', ', $sets) 
+                        . ' WHERE id = :id AND user_id = :user_id', self::TASKS_TABLE);
 
             $stmt = $this->connection->prepare($sql);
 
-            $stmt->bindValue('id', $id, PDO::PARAM_INT);
+            $stmt->bindValue(':id', $id, PDO::PARAM_INT);
 
             foreach ($fields as $name => $values) {
                 $stmt->bindValue($name, $values[0], $values[1]);
             }
+
+            $stmt->bindValue(':user_id', $userId, PDO::PARAM_INT);
 
             $stmt->execute();
 
@@ -99,13 +108,14 @@ class TaskGateway
         }
     }
 
-    public function delete(string $id): int
+    public function delete(int $userId, string $id): int
     {
-        $sql = sprintf('DELETE FROM %s WHERE id = :id', self::TASKS_TABLE);
+        $sql = sprintf('DELETE FROM %s WHERE id = :id AND user_id = :user_id', self::TASKS_TABLE);
 
         $stmt = $this->connection->prepare($sql);
 
-        $stmt->bindValue('id', $id, PDO::PARAM_INT);
+        $stmt->bindValue(':id', $id, PDO::PARAM_INT);
+        $stmt->bindValue(':user_id', $userId, PDO::PARAM_INT);
 
         $stmt->execute();
